+++ aider/io.py
@@ -192,6 +192,7 @@ class InputOutput:
         dry_run=False,
         llm_history_file=None,
         editingmode=EditingMode.EMACS,
+        fancy_input=True,
     ):
         self.never_prompts = set()
         self.editingmode = editingmode
@@ -234,7 +235,7 @@ class InputOutput:
         self.append_chat_history(f"\n# aider chat started at {current_time}\n\n")
 
         self.prompt_session = None
-        if self.pretty:
+        if fancy_input:
             # Initialize PromptSession
             session_kwargs = {
                 "input": self.input,
@@ -359,7 +360,7 @@ class InputOutput:
         show = ""
         if rel_fnames:
             rel_read_only_fnames = [
-                os.path.relpath(fname, root) for fname in (abs_read_only_fnames or [])
+                get_rel_fname(fname, root) for fname in (abs_read_only_fnames or [])
             ]
             show = self.format_files_for_input(rel_fnames, rel_read_only_fnames)
         if edit_format:
@@ -701,13 +702,20 @@ class InputOutput:
 
     def format_files_for_input(self, rel_fnames, rel_read_only_fnames):
         read_only_files = []
-        for full_path in rel_read_only_fnames or []:
+        for full_path in sorted(rel_read_only_fnames or []):
             read_only_files.append(f"{full_path} (read only)")
 
         editable_files = []
-        for full_path in rel_fnames:
+        for full_path in sorted(rel_fnames):
             if full_path in rel_read_only_fnames:
                 continue
             editable_files.append(f"{full_path}")
 
         return "\n".join(read_only_files + editable_files) + "\n"
+
+
+def get_rel_fname(fname, root):
+    try:
+        return os.path.relpath(fname, root)
+    except ValueError:
+        return fname